backend/requirements.txt
# ── Web Framework ──────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Database ───────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# ── Redis / Celery ─────────────────────────────────────────
redis==5.0.4
celery==5.3.6
kombu==5.3.4

# ── Auth ───────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# ── Validation ─────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# ── HTTP ───────────────────────────────────────────────────
httpx==0.27.0
aiohttp==3.9.5
tenacity==8.2.3

# ── AI / LLM ───────────────────────────────────────────────
google-generativeai==0.5.4
langchain-core==0.2.9
langgraph==0.1.19

# ── Vector DB ──────────────────────────────────────────────
chromadb==0.5.0

# ── Resume Parsing ─────────────────────────────────────────
pdfminer.six==20231228
pymupdf==1.24.3
spacy==3.7.4

# ── Utils ──────────────────────────────────────────────────
python-slugify==8.0.4
python-dateutil==2.9.0.post0
